Separated experimental curves to new file
